Revert "fixed server segfault maybe"
This reverts commit 3810944a238ea154ecfb81dd7b7c246375271256.

Run git pull to get non-screwed up indentation... I'll apply the change
you made (moving the line that removes the player at the end of the
function) with a non-screwed up file.

Trying to change the style (especially indentation) of programs is a bad
idea... it's much better to decide well in the beginning and stick with
that.