Refactored some of the code in the client and added a few comments.
Also, removed the code to accept players if the player count ever goes
below the minimum -- if we ever are to accept additional players after the
game has started, it should probably be up to the maximum number of
players. Or infinity, since we have linked lists stored in the heap, but
checking calloc for errors? :p