* Switched to 3-space indentation (4 is better but gnaroengaeognag) in the   srv_player.c so that it matches the rest of the files. * Fixed bug with linked list in the server, and added more debug prints. * Moved some linked-list specific functions from server.c to srv_player.c
Still need to fix the linked list on the client, and make the server quit
or reset when there is only one player left, and perhaps get rid of the
flood of debugging prints.

Should send PLAYER_WIN to clients if they reach the exit, probably. And
add another exit in the opposite side of the randomly generated maze to
prevent the predator from blocking it.